+++ app/Plugin/Admin/Admin.php
@@ -348,6 +348,13 @@ class Admin {
 			$transient_obj->set_message( __( '<strong>The running import has been canceled.</strong> Click on the following button to start a new import. If it also takes to long please check your hosting logfiles for possible restrictions mentioned there.', 'personio-integration-light' ) . ' <br><br><a href="' . esc_url( Helper::get_import_url() ) . '" class="button button-primary personio-integration-import-hint">' . __( 'Run import', 'personio-integration-light' ) . '</a>' );
 			$transient_obj->set_type( 'error' );
 			$transient_obj->save();
+
+			// get active user.
+			$user = wp_get_current_user();
+
+			// log this event.
+			$log = new Log();
+			$log->add_log( sprintf( __( 'A running import has been canceled through %1$s.', 'personio-integration-light' ), esc_html( $user->display_name ) ), 'info', 'import' );
 		}
 
 		// redirect user.